fix win failure of TestDocComments

+++ test/testng/org/openjdk/jextract/test/toolprovider/Test7903257/typedefs.h
@@ -21,7 +21,7 @@
  * questions.
  */
 
-typedef unsigned long size_t;
+typedef unsigned long long size_t;
 typedef int INT_32;
 typedef int* INT_PTR;
 typedef struct Foo* OPAQUE_PTR;
